# Vectorise random vectors of `Float16`