--- a/stdlib/Random/src/XoshiroSimd.jl
+++ b/stdlib/Random/src/XoshiroSimd.jl
@@ -45,11 +45,15 @@ simdThreshold(::Type{Bool}) = 640
     (UInt64(reinterpret(UInt32, u)) << 32) | UInt64(reinterpret(UInt32, l))
 end
 @inline function _bits2float(x::UInt64, ::Type{Float16})
-    ui = (x>>>16) % UInt16
-    li = x % UInt16
-    u = Float16(ui >>> 5) * Float16(0x1.0p-11)
-    l = Float16(li >>> 5) * Float16(0x1.0p-11)
-    (UInt64(reinterpret(UInt16, u)) << 16) | UInt64(reinterpret(UInt16, l))
+    i1 = (x>>>48) % UInt16
+    i2 = (x>>>32) % UInt16
+    i3 = (x>>>16) % UInt16
+    i4 = x % UInt16
+    f1 = Float16(i1 >>> 5) * Float16(0x1.0p-11)
+    f2 = Float16(i2 >>> 5) * Float16(0x1.0p-11)
+    f3 = Float16(i3 >>> 5) * Float16(0x1.0p-11)
+    f4 = Float16(i4 >>> 5) * Float16(0x1.0p-11)
+    return (UInt64(reinterpret(UInt16, f1)) << 48) | (UInt64(reinterpret(UInt16, f2)) << 32) | (UInt64(reinterpret(UInt16, f3)) << 16) | UInt64(reinterpret(UInt16, f4))
 end
 
 # required operations. These could be written more concisely with `ntuple`, but the compiler