With this PR, `rand!(::Vector{Float16})` on Apple Silicon becomes about one order of magnitude faster:
```
julia> @btime rand!(v) setup=(v=Vector{Float16}(undef, 2 ^ 20));
  1.208 ms (0 allocations: 0 bytes) # before PR
  122.750 μs (0 allocations: 0 bytes) # after PR
```
and about 2x and 4x faster than vectors of `Float32` or `Float64` of the same size, when the operation is memory-bound, as you'd expect:
```
julia> @btime rand!(v) setup=(v=Vector{Float32}(undef, 2 ^ 20));
  375.125 μs (0 allocations: 0 bytes)

julia> @btime rand!(v) setup=(v=Vector{Float64}(undef, 2 ^ 20));
  763.667 μs (0 allocations: 0 bytes)
```

This sensibly improves performance also on hardware without native `Float16`, for example on zenver2 (AMD Ryzen 9 3950X) I get
```
julia> @btime rand!(v) setup=(v=Vector{Float16}(undef, 2 ^ 20));
  1.180 ms (0 allocations: 0 bytes) # before PR
  335.613 μs (0 allocations: 0 bytes) # after PR
```

<s>However I'm pretty sure I did something wrong because now `rand!(::Vector{Float16})` produces numbers larger than 1, so I need help to fix my mindless pattern matching adaptation 🙂 My assumption is that I got wrong the implementation of `_bits2float`, which I have little clue of what it's supposed to do exactly (besides converting a `UInt64` to a floating point number) being undocumented and sparsely commented.  CC @chethega who I believe originally introduced this code in #34852/#40546.  Also CC @rfourquet in case he can provide some help.</s>  ***Edit***: wait, I think I got it 👀