@oscardssmith 2024-10-05
Looks great! Any chance you can add similar code to BFloat16.jl?