@giordano 2024-10-05
> Any chance you can add similar code to BFloat16.jl?

_Maybe_