@giordano 2024-10-05
Argh, I just realised there's a problem: while statistical properties of the generated numbers are good:
```julia-repl
julia> using Statistics

julia> v = Float32.(rand(Float16, 100)); mean(v), std(v)
(0.49073243f0, 0.28596097f0)

julia> v = Float32.(rand(Float16, 10000)); mean(v), std(v)
(0.5040806f0, 0.28860313f0)

julia> v = Float32.(rand(Float16, 1000000)); mean(v), std(v)
(0.49959266f0, 0.28868693f0)
```
the uniform distribution has mean $\approx 0.5$ and standard deviation $\approx\sqrt{1/12}$ as expected, the stream of numbers returns 0 every two numbers:
```julia-repl
julia> rand(Float16, 16)'
1×16 adjoint(::Vector{Float16}) with eltype Float16:
 0.0674  0.145  0.0  0.0  0.1396  0.617  0.0  0.0  0.653  0.856  0.0  0.0  0.713  0.937  0.0  0.0

julia> rand(Float16, 16)'
1×16 adjoint(::Vector{Float16}) with eltype Float16:
 0.8984  0.881  0.0  0.0  0.1489  0.4072  0.0  0.0  0.03564  0.4263  0.0  0.0  0.1499  0.1758  0.0  0.0

julia> rand(Float16, 16)'
1×16 adjoint(::Vector{Float16}) with eltype Float16:
 0.2373  0.0913  0.0  0.0  0.1978  0.771  0.0  0.0  0.29  0.833  0.0  0.0  0.3145  0.5767  0.0  0.0

julia> rand(Float16, 16)'
1×16 adjoint(::Vector{Float16}) with eltype Float16:
 0.23  0.3013  0.0  0.0  0.168  0.5425  0.0  0.0  0.9404  0.006348  0.0  0.0  0.2603  0.371  0.0  0.0
```
I guess there's something wrong in the `llvmcall` 🥲 If anyone can spot the issue before, please let me know!